Generated by Spark: put a validation on the schedule form to make sure the end date is at least 30 minutes past the start date, as a default put an end date of start date +90 minutes.
the request cancellation button should be aligned to the event type (game, practice, meeting, other) and the status label. it should be positioned right to type and status. make the labels less colorful and leave a bit less space around it, particularly on the top. the card padding can be significantly reduced at the top.