Generated by Spark: in operations office for sites, allow an active/inactive toggle. if a site is set to inactive, all related fields are not available for selection. also include fields to store coordinates and a link to the site pointing to maps.google.ch.
in fields, remove the sport type for field as it is not necessary.

when creating a new event, make sure that if the event type is game, there can only be a single team selected. for all events created, automatically add head coach and team manager in the dashboard view.