Generated by Spark: i want to add a cancellation button for all events  in the dashboard view that are more than 36 hours in the future. the cancellation button should be visually aligned to the event type and event status labels, but more to the right. use a similar design, and label the button 'Request Cancellation'. The button will trigger a form where a user must provide a name and a phone number along with a justification to cancel a practice. The request will be shown in the request section in Operations Office. Also, the request cancellation button can only be shown on events that are either confirmed or planned. Also, switch any event that is less than 24 hours in the future to status confirmed if the status is planned. there should not be an event in status planned with a start date in less than 24 hours.
Furthermore, please adjust field/button size of the teams dropdown, request facility, request equipment, operations office and the list/schedule toggle so that the full width of the site is used and the buttons should be similar in size.

also make sure that all icons are similar in size, line thickness, stile. for larger icons, use a bit of shade, smaller icons need to be readable.

for the admin user use a crown as an icon, for the mgmt user a badge.

in the operations office view, remove the 'dashboard' button link in the top bar next to the user.