Generated by Spark: make the avatar for users a line drawing in a white circle - like an icon but not colorful.
i would prefer the toggles to have a background so it looks more like a slider. i like the bright color, but would prefer the color to be more in line with the overall page design, no purple or green. maybe the background should be the more colorful element with the selection being bright.